requirement.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
cycler==0.10.0
Django==2.1
djangorestframework==3.10.2
joblib==0.13.2
kiwisolver==1.1.0
kombu==3.0.37
lightgbm==2.2.3
matplotlib==3.1.1
netifaces==0.10.9
numpy==1.17.0
pandas==0.25.0
psutil==5.6.3
psycopg2==2.7.7
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.0
six==1.14.0
